Simplify view methods by forwarding to methods of the controller
We had quite some code duplication in the view methods, what we resolve
by forwarding to methods of the respective controllers. This requires
$this to be bound to the closures, which is only available as of PHP
5.4.0.

An alternative might have been to `use` $this and to make the
respective controller methods public, but this would be an ugly hack,
which would break as of PHP 7.1.0. Anyway, there shouldn't be a real
problem requiring PHP 5.4.0 in 2017.