Unify README and help files
README.md is particularly useful for Github, where full-fledged usage
information is desirable.  Also it is more convenient to write Markdown
than HTML.  Therefore we make README.md the canonical format, and
generate help.htm via Pandoc.

Sadly, that requires to remove the Czech manual, since that was already
out-dated, and we don't speak that language.

We also remove the READMEs from the distribution builds, since
CMSimple_XH users are accustomed to find the manuals in help/.