Fix #3: Unpublished pages are not ignored
We just check whether the `published` flag in the page data is unset.
That does not cater to pages which are `#cmsimple remove#`d (what we
just ignore, because that shouldn't used be anyway), and it doesn't
catch dynamic changes of plugins which are loaded after Yanp.  We could
move the check to after_plugin_loading, but that wouldn't necessarily
cut it either.  Well, "y" is late in the alphabet, so maybe not an
issue in practice.